% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots a state map of accidents for a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer representing the state number for which
the accident map will be generated}

\item{year}{input value of the year for which we'll plot the
accidents for the selected state}
}
\value{
A plot of all accident locations rendered on the map of
  the input state. A \code{invalid STATE number} message will be
  displayed if \code{state.num} input value is not found in the
  input file. If no accidents happened in that state during that
  calendar year, a \code{no accidents to plot} will be displayed
  instead of the plot.
}
\description{
\code{fars_map_state} Plots a state map with all the accident
  locations that had happened during the input year. All
  longitudinal and latitudinal values are filtered to make sure
  they fall within standard intervals.
}
\examples{
\dontrun{fars_map_state(1,2014)}
\dontrun{fars_map_state(3,2015)}

}
