% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a delimited data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string prividing the name of the file or
the path to a file.}
}
\value{
If the file is found then the output will be a data frame object
  containing the content of the file. If the file does not exist an
  error message will be printed stating that the file does not exist.
}
\description{
\code{fars_read} reads a delimited csv file and
convert the result to a data frame.
}
\examples{
\dontrun{fars_read('accident_2013.csv.bz2')}
\dontrun{fars_read('/var/tmp/accident_2014.csv.bz2')}

}
