% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads month and year values from accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of years to get data for}
}
\value{
list of tibbles, one per year, with two columns
  MONTH and year. If any of the years from the input vector
  does not corresponds to an existing data file for that
  particular year, NULL and \code{invalid year} warning message
  will be returned
}
\description{
\code{fars_read_years} reads multiple yearly data files
  and returns a list of tibbles with MONTH and year
  columns from the data in "accident_<year>.csv.bz2"
  files
}
\examples{
\dontrun{fars_read_years('2015')}
\dontrun{fars_read_years(c('2013','2014'))}
}
