% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the number of accidents/month for each year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years to get accidents/month summary data for}
}
\value{
a matrix containing 12 rows (for each month) and one column
  for each year with the total number of accidents for that
  particular month in that particular year. A warning is returned
  for each year that does not have a corresponding "accident_<year>.csv.bz2"
  file.
}
\description{
\code{fars_summarize_years} diplays a summary of the number of US
  accidents per month for each year
}
\examples{
\dontrun{fars_summarize_years(2015)}
\dontrun{fars_summarize_years(c(2013,2014))}

}
