% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The input value for the year}
}
\value{
This function returns a string in the form
 "accident_<year>.csv.bz2" that can be used as a file
 name that contains the data for that particular \code{year}
}
\description{
\code{make_filename} generates the name of an input data file
  based on the \code{year} argument provided in a form
  "accident_<year>.csv.bz2"
}
