% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_season_form.R
\name{plot_season_form}
\alias{plot_season_form}
\title{Plot Season Form}
\usage{
plot_season_form(league, season_start_year)
}
\arguments{
\item{league}{name of league (example: "EPL")}

\item{season_start_year}{season start year}
}
\value{
plot with form of teams in specified season
}
\description{
Plot Season Form
}
\examples{
plot_season_form("EPL", 2022)
}
